Changed dimensions for more flexibility on the speed
The previous distance between the x and y coordinates of two adjacent squares was 110, which has only two prime dividers, namely 2 and 55. This limits the value of the speed to one of those two values in my current implementation, which is why I decided to change the dimensions so that now the distance between two adjacent squares is 120, which has many more dividers and therefore provides much more flexibility.